API test: move digest functions out
Move all api.c tests of wolfCrypt APIs that are for digests out into
separate files.
