+++ tests/api/include.am
@@ -1,5 +1,30 @@
+# vim:ft=automake
+# included from Top Level Makefile.am
+# All paths should be given relative to the
+
 if BUILD_TESTS
-tests_unit_test_SOURCES += tests/api/ascon.c
+tests_unit_test_SOURCES += tests/api/test_md5.c
+tests_unit_test_SOURCES += tests/api/test_sha.c
+tests_unit_test_SOURCES += tests/api/test_sha256.c
+tests_unit_test_SOURCES += tests/api/test_sha512.c
+tests_unit_test_SOURCES += tests/api/test_sha3.c
+tests_unit_test_SOURCES += tests/api/test_blake2.c
+tests_unit_test_SOURCES += tests/api/test_sm3.c
+tests_unit_test_SOURCES += tests/api/test_ripemd.c
+tests_unit_test_SOURCES += tests/api/test_hash.c
+tests_unit_test_SOURCES += tests/api/test_ascon.c
 endif
-EXTRA_DIST += tests/api/ascon.h
-EXTRA_DIST += tests/api/ascon_kats.h
+EXTRA_DIST += tests/api/api.h
+EXTRA_DIST += tests/api/test_md5.h
+EXTRA_DIST += tests/api/test_sha.h
+EXTRA_DIST += tests/api/test_sha256.h
+EXTRA_DIST += tests/api/test_sha512.h
+EXTRA_DIST += tests/api/test_sha3.h
+EXTRA_DIST += tests/api/test_blake2.h
+EXTRA_DIST += tests/api/test_sm3.h
+EXTRA_DIST += tests/api/test_ripemd.h
+EXTRA_DIST += tests/api/test_hash.h
+EXTRA_DIST += tests/api/test_ascon.h
+EXTRA_DIST += tests/api/test_ascon.h
+EXTRA_DIST += tests/api/test_ascon_kats.h
+
